Change argument csv to list
The semantics of "csv" as an argument makes less sense, now that
we create also the Excel-file. What we are really doing is converting
a very long row-based format into a list or table-format. So, we
will use the term "list" so indicate the operation we're doing.

Refactored the tests also a bit, coverage got increased to 99%.