README.md - pytest xml coverage & .gitignore
SonarQube wants the pytest coverage, so that it can report it
correctly. This should allow SonarQube to find what it expects
from the default locations it has.
- You do have to set the Python language level to 3.6.
- This is done to:

```
<installation_path>/sonar-scanner-<version>-linux/conf/sonar-scanner.properties
```

Add the row:

```
sonar.python.version=3.6
``

SonarQube cache and also the expected coverage/test result locations
should be ignored by default.

Seems SonarQube can only pick up the coverage if they are located
in the root of the repo. Quite mysterious, could be related to:

https://github.com/nedbat/coveragepy/issues/578